chore: reorganize signup validators
